Move timer.py from libs to utils
